@@ -3,12 +3,12 @@ package com.myhitchhikingspots;
 import android.content.Context;
 import android.graphics.PixelFormat;
 import android.os.IBinder;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.design.widget.CoordinatorLayout;
-import android.support.design.widget.Snackbar;
-import android.support.v4.view.GravityCompat;
-import android.support.v4.view.ViewCompat;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.coordinatorlayout.widget.CoordinatorLayout;
+import com.google.android.material.snackbar.Snackbar;
+import androidx.core.view.GravityCompat;
+import androidx.core.view.ViewCompat;
 import android.view.ContextThemeWrapper;
 import android.view.Gravity;
 import android.view.View;
