@@ -16,22 +16,22 @@ import android.location.Address;
 import android.os.AsyncTask;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.design.internal.BottomNavigationItemView;
-import android.support.design.widget.BottomNavigationView;
-import android.support.design.widget.BottomSheetBehavior;
-import android.support.design.widget.CoordinatorLayout;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.v4.content.ContextCompat;
-import android.support.v4.view.GravityCompat;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v4.widget.NestedScrollView;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.app.AppCompatDelegate;
-import android.support.v7.widget.AppCompatImageButton;
-import android.support.v7.widget.Toolbar;
+import androidx.annotation.NonNull;
+import com.google.android.material.bottomnavigation.BottomNavigationItemView;
+import com.google.android.material.bottomnavigation.BottomNavigationView;
+import com.google.android.material.bottomsheet.BottomSheetBehavior;
+import androidx.coordinatorlayout.widget.CoordinatorLayout;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+import com.google.android.material.snackbar.Snackbar;
+import androidx.core.content.ContextCompat;
+import androidx.core.view.GravityCompat;
+import androidx.drawerlayout.widget.DrawerLayout;
+import androidx.core.widget.NestedScrollView;
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatDelegate;
+import androidx.appcompat.widget.AppCompatImageButton;
+import androidx.appcompat.widget.Toolbar;
 import android.text.TextUtils;
 import android.util.Log;
 import android.view.Display;
@@ -182,7 +182,7 @@ public class SpotFormActivity extends AppCompatActivity implements RatingBar.OnR
     private ImageView dropPinView;
 
     private CoordinatorLayout coordinatorLayout, spot_form_basic;
-    private android.support.design.widget.FloatingActionButton fabLocateUser, fabZoomIn, fabZoomOut;
+    private FloatingActionButton fabLocateUser, fabZoomIn, fabZoomOut;
 
     private NestedScrollView scrollView;
     BottomSheetBehavior mBottomSheetBehavior;
@@ -1900,7 +1900,7 @@ public class SpotFormActivity extends AppCompatActivity implements RatingBar.OnR
         snackbar.setActionTextColor(Color.BLACK);
 
         // change snackbar text color
-        int snackbarTextId = android.support.design.R.id.snackbar_text;
+        int snackbarTextId = com.google.android.material.R.id.snackbar_text;
         TextView textView = (TextView) snackbarView.findViewById(snackbarTextId);
         if (textView != null) textView.setTextColor(Color.WHITE);
 
@@ -2195,7 +2195,7 @@ public class SpotFormActivity extends AppCompatActivity implements RatingBar.OnR
                                 .parseDateTime(placeWithCompleteDetails.getDescriptionENdatetime());
                         mCurrentSpot.setStartDateTime(extractedDateTime);
                     }
-                    
+
                     //If a waiting time was informed, convert it into Integer
                     String waiting_stats_avg = placeWithCompleteDetails.getWaiting_stats_avg();
                     if (waiting_stats_avg != null && !waiting_stats_avg.equalsIgnoreCase("null") && !waiting_stats_avg.isEmpty())
